refactor: split app profile management helpers
